# backend/internal/db/sqlc/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package dbsqlc

import (
	"github.com/jackc/pgx/v5/pgtype"
)

type Message struct {
	ID        int64              `json:"id"`
	SessionID int64              `json:"session_id"`
	Role      string             `json:"role"`
	Content   string             `json:"content"`
	Metadata  []byte             `json:"metadata"`
	CreatedAt pgtype.Timestamptz `json:"created_at"`
}

type ModelProfile struct {
	ID              string             `json:"id"`
	Name            string             `json:"name"`
	Provider        string             `json:"provider"`
	ModelID         string             `json:"model_id"`
	BaseUrl         string             `json:"base_url"`
	ApiKeyEnv       string             `json:"api_key_env"`
	ContextWindow   int32              `json:"context_window"`
	MaxOutputTokens int32              `json:"max_output_tokens"`
	Temperature     float64            `json:"temperature"`
	TimeoutSeconds  int32              `json:"timeout_seconds"`
	Status          string             `json:"status"`
	Metadata        []byte             `json:"metadata"`
	CreatedAt       pgtype.Timestamptz `json:"created_at"`
	UpdatedAt       pgtype.Timestamptz `json:"updated_at"`
	DeletedAt       pgtype.Timestamptz `json:"deleted_at"`
	ApiKey          string             `json:"api_key"`
}

type Project struct {
	ID          string             `json:"id"`
	Name        string             `json:"name"`
	Description string             `json:"description"`
	Status      string             `json:"status"`
	Metadata    []byte             `json:"metadata"`
	CreatedAt   pgtype.Timestamptz `json:"created_at"`
	UpdatedAt   pgtype.Timestamptz `json:"updated_at"`
	DeletedAt   pgtype.Timestamptz `json:"deleted_at"`
}

type ProjectDocument struct {
	ID        int64              `json:"id"`
	ProjectID string             `json:"project_id"`
	Kind      string             `json:"kind"`
	Title     string             `json:"title"`
	Body      string             `json:"body"`
	Metadata  []byte             `json:"metadata"`
	CreatedAt pgtype.Timestamptz `json:"created_at"`
	UpdatedAt pgtype.Timestamptz `json:"updated_at"`
	DeletedAt pgtype.Timestamptz `json:"deleted_at"`
}

type Run struct {
	ID        int64              `json:"id"`
	ProjectID pgtype.Text        `json:"project_id"`
	SessionID pgtype.Int8        `json:"session_id"`
	Input     string             `json:"input"`
	FinalText string             `json:"final_text"`
	RunDir    string             `json:"run_dir"`
	Status    string             `json:"status"`
	Error     string             `json:"error"`
	Metadata  []byte             `json:"metadata"`
	CreatedAt pgtype.Timestamptz `json:"created_at"`
	UpdatedAt pgtype.Timestamptz `json:"updated_at"`
}

type Session struct {
	ID        int64              `json:"id"`
	ProjectID pgtype.Text        `json:"project_id"`
	Title     string             `json:"title"`
	Status    string             `json:"status"`
	Metadata  []byte             `json:"metadata"`
	CreatedAt pgtype.Timestamptz `json:"created_at"`
	UpdatedAt pgtype.Timestamptz `json:"updated_at"`
	DeletedAt pgtype.Timestamptz `json:"deleted_at"`
}